modelapp/requirements.txt
numpy
flask
flask_cors
Pillow
opencv-python
fer
pymongo
pandas
tensorflow
gunicorn